Improved UI and cleaned up folders
