+++ Tower Defence/Assets/Prefabs/UpgradeSelectionOption.prefab
@@ -140,7 +140,9 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_text: Turret Evolution
+  m_text: 'Turret Upgrade
+
+'
   m_isRightToLeft: 0
   m_fontAsset: {fileID: 11400000, guid: ed862fa2de5a148a8adb187ac55cd3eb, type: 2}
   m_sharedMaterial: {fileID: 7459919616753430026, guid: ed862fa2de5a148a8adb187ac55cd3eb, type: 2}
@@ -386,7 +388,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_text: 'Can be applied to:'
+  m_text: 'Applicable Turrets:'
   m_isRightToLeft: 0
   m_fontAsset: {fileID: 11400000, guid: ed862fa2de5a148a8adb187ac55cd3eb, type: 2}
   m_sharedMaterial: {fileID: 7459919616753430026, guid: ed862fa2de5a148a8adb187ac55cd3eb, type: 2}
@@ -737,7 +739,7 @@ GameObject:
   - component: {fileID: 308355810}
   - component: {fileID: 912509167598602563}
   m_Layer: 5
-  m_Name: EvolutionSelectionOption
+  m_Name: UpgradeSelectionOption
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -861,6 +863,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 942cf2bd308b34e158307b0e9816254c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  upgrade: {fileID: 0}
   iconImage: {fileID: 4543250165993662423}
   displayName: {fileID: 8037839850094324763}
   effect: {fileID: 1219808412639005329}
