--- Tower Defence/Assets/Scripts/Turrets/Upgrades/BulletUpgrades.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 167ce6e0d3f704e17a8f7083c18a2888
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
