+++ Tower Defence/Assets/Scripts/Turrets/Upgrades/DamageUpgrade.cs
@@ -1,9 +1,10 @@
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace Turrets.Upgrades.BulletUpgrades
 {
-    [CreateAssetMenu(fileName = "MissileBulletT0", menuName = "Upgrades/MissileBullet")]
-    public class MissileBullet : Upgrade
+    [CreateAssetMenu(fileName = "BulletDamageT0", menuName = "Upgrades/BulletDamage")]
+    public class DamageUpgrade : Upgrade
     {
         public override void AddUpgrade(Turret turret)
         {
@@ -16,10 +17,10 @@ namespace Turrets.Upgrades.BulletUpgrades
         }
 
         public override void OnShoot(Bullet bullet)
-        {;
-            bullet.explosionRadius += GETUpgradeValue();
+        {
+            bullet.damage += (int) (GETUpgradeValue() * bullet.damage);
         }
 
-        public override void OnHit(Enemy[] targets) { }
+        public override void OnHit(IEnumerable<Enemy> targets) { }
     }
 }
