--- Tower Defence/Assets/Scripts/Turrets/Upgrades/TurretUpgrades.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 391c2ffd841a54f80b5d6ead3ae1642c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
